updated prompts/classes and added generate_sql function
