+++ pip_library_etl/__init__.py
@@ -1,0 +1,7 @@
+from pip_library_etl.main import (
+    PipLibraryEtl
+)
+
+__all__ = [
+    "PipLibraryEtl"
+]
