--- pip_library_parser/__init__.py
@@ -1,7 +1,0 @@
-from pip_library_parser.main import (
-    CodeToDocGenerator
-)
-
-__all__ = [
-    "CodeToDocGenerator"
-]
